requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.9.0
attrs==25.1.0
azure-ai-documentintelligence==1.0.2
azure-core==1.33.0
azure-identity==1.21.0
beautifulsoup4==4.13.3
black==24.2.0
blessed==1.20.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
coloredlogs==15.0.1
colorlog==6.9.0
cryptography==44.0.2
debugpy==1.8.12
defusedxml==0.7.1
distro==1.9.0
editor==1.6.6
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.7.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
inquirer==3.4.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
lxml==5.3.2
magika==0.6.1
mammoth==1.9.0
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.6.1
markitdown==0.1.1
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.4
olefile==0.47
onnxruntime==1.21.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20250327
pdftext==0.6.2
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
PyMuPDF==1.25.4
pyparsing==3.2.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
ruff==0.3.0
runs==1.2.2
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==76.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
starlette==0.46.2
surya-ocr==0.13.0
sympy==1.13.1
temporalio==1.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
types-protobuf==5.29.1.20250208
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
wcwidth==0.2.13
websockets==14.2
xlrd==2.0.1
XlsxWriter==3.2.2
xmod==1.8.1
yarl==1.18.3
youtube-transcript-api==1.0.3
